Add Tekton bits for running in Kubernetes
The container base is fedora because that enables installing the python
dependencies via RPMs. It is specifically `34` because I had some DNS
issues with newer versions.

The chgrp in the Containerfile is to enable it to work on OpenShift,
where it can be run with an (effectively) random user ID.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>